fix: 의상API 파일명 변경, post 방식 수정 파라미터 추가
